+++ cmd/pricing-cli/main_test.go
@@ -4,13 +4,13 @@ import (
 	"os"
 	"testing"
 
-	chassis "github.com/ai8future/chassis-go/v6"
-	"github.com/ai8future/chassis-go/v6/secval"
-	"github.com/ai8future/chassis-go/v6/testkit"
+	chassis "github.com/ai8future/chassis-go/v8"
+	"github.com/ai8future/chassis-go/v8/secval"
+	"github.com/ai8future/chassis-go/v8/testkit"
 )
 
 func TestMain(m *testing.M) {
-	chassis.RequireMajor(6)
+	chassis.RequireMajor(8)
 	os.Exit(m.Run())
 }